Isolate view from page presenter file
