+++ Core/Core/Common/CommonUI/Presenter/PagingPresenter.swift
@@ -78,83 +78,3 @@ public class PagingPresenter<Controller: PagingViewController> {
         controller.loadNextPage()
     }
 }
-
-public class PageLoadingCell: UITableViewCell {
-
-    private let progressView = CircleProgressView()
-    private let label = UILabel()
-    private var subscriptions = Set<AnyCancellable>()
-
-    override public init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-
-        contentView.addSubview(progressView)
-
-        progressView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            progressView.widthAnchor.constraint(equalToConstant: 32),
-            progressView.heightAnchor.constraint(equalToConstant: 32),
-            progressView.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
-            progressView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 10),
-            progressView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -10)
-        ])
-
-        contentView.addSubview(label)
-
-        label.text = String(localized: "Load More", bundle: .core)
-        label.textColor = .systemBlue
-        label.font = .scaledNamedFont(.semibold16)
-        label.adjustsFontForContentSizeCategory = true
-        label.translatesAutoresizingMaskIntoConstraints = false
-
-        NSLayoutConstraint.activate([
-            label.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
-            label.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-            label.topAnchor.constraint(greaterThanOrEqualTo: contentView.topAnchor, constant: 10),
-            label.bottomAnchor.constraint(lessThanOrEqualTo: contentView.bottomAnchor, constant: -10)
-        ])
-    }
-
-    required init?(coder: NSCoder) { nil }
-
-    override public func layoutSubviews() {
-        super.layoutSubviews()
-
-        subviews.forEach { subview in
-            guard subview != contentView else { return }
-            subview.isHidden = true
-        }
-    }
-
-    override public func prepareForReuse() {
-        super.prepareForReuse()
-
-        subscriptions.forEach({ $0.cancel() })
-        subscriptions.removeAll()
-    }
-
-    fileprivate func observeLoading(_ loadingPublisher: AnyPublisher<Bool, Never>) {
-        loadingPublisher
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] isLoading in
-                if isLoading {
-                    self?.setupAsProgressView()
-                } else {
-                    self?.setupAsButton()
-                }
-            }
-            .store(in: &subscriptions)
-    }
-
-    private func setupAsButton() {
-        backgroundConfiguration = UIBackgroundConfiguration.listPlainCell()
-        progressView.isHidden = true
-        label.isHidden = false
-    }
-
-    private func setupAsProgressView() {
-        backgroundConfiguration = UIBackgroundConfiguration.clear()
-        progressView.isHidden = false
-        label.isHidden = true
-    }
-}
